Create Achievements
Simply duplicate the cells to create new missions.

Pay attention to the commas: The last element should never get a comma.

Title and descriptions can be left blank. They will be generated automatically depending on the mission type and data. It is recommended that you write them for special missions. Example: The "First Mile" mission has a custom title and a description. If it didn't have them, the system would use: "1 Mile" for the title and "Travel 1 mile" for the description.

1 Mile ~ 5750 Roblox Studs.